+++ src/index/bits.rs
@@ -1,10 +1,7 @@
 //! Bit twiddling.
 
 use super::IndexMode;
-use crate::{
-    coord::CoordIJK, Direction, Edge, Resolution, Vertex, CCW, CW,
-    DIRECTION_BITSIZE,
-};
+use crate::{Direction, Edge, Resolution, Vertex, DIRECTION_BITSIZE};
 use std::{cmp, num::NonZeroU8};
 
 /// Offset (in bits) of the mode in an H3 index.
@@ -249,41 +246,6 @@ pub fn pentagon_rotate60<const CCW: bool>(mut bits: u64) -> u64 {
     }
 
     bits
-}
-
-/// Set the directions of a cell index (in-place) from finest resolution up.
-///
-/// IJK coordinates are adjusted during the traversal so that, at the end, they
-/// should match the IJK of the base cell in the coordinate system of the
-/// current base cell.
-///
-/// Returns the adjusted `IJK` coordinates.
-#[allow(clippy::inline_always)] // 4-5% boost, up to 13% at resolution 1.
-#[inline(always)]
-pub fn directions_bits_from_ijk(
-    mut ijk: CoordIJK,
-    bits: &mut u64,
-    resolution: Resolution,
-) -> CoordIJK {
-    for res in Resolution::range(Resolution::One, resolution).rev() {
-        let last_ijk = ijk;
-        let last_center = if res.is_class3() {
-            // Rotate CCW.
-            ijk = ijk.up_aperture7::<{ CCW }>();
-            ijk.down_aperture7::<{ CCW }>()
-        } else {
-            // Rotate CW.
-            ijk = ijk.up_aperture7::<{ CW }>();
-            ijk.down_aperture7::<{ CW }>()
-        };
-
-        let diff = (last_ijk - last_center).normalize();
-        let direction = Direction::try_from(diff).expect("unit IJK coordinate");
-        // SAFETY: `res` is in [resolution; 1], thus valid.
-        *bits = set_direction(*bits, direction.into(), res);
-    }
-
-    ijk
 }
 
 #[cfg(test)]